add test numbers

+++ app/Http/Resources/OfferResource.php
@@ -14,7 +14,7 @@ class OfferResource extends JsonResource
             'title'                 => $this->title,
             'description'           => $this->description,
             'terms'                 => $this->terms,
-            'price'                 => (float)$this->price,
+            'price'                 => (int)$this->price,
             'discount'              => $this->discount,
             'discount_string'       => $this->discount_string,
             'status'                => $this->status,
